requirements.txt
pyqt5
qscintilla
psycopg2
psutil
cryptography
pysnmp
redispy